requirements.txt
jittor==1.3.0.2
numpy==1.17.4
opencv_python==4.4.0.46
Pillow==9.0.0
tqdm==4.54.1